+++ client/web/src/search/results/sidebar/SearchReference.tsx
@@ -611,8 +611,9 @@ export interface SearchReferenceProps
 const SearchReference = (props: SearchReferenceProps): ReactElement => {
     const [selectedTab, setSelectedTab] = useLocalStorage(SEARCH_REFERENCE_TAB_KEY, 0)
 
-    const { onNavbarQueryChange, navbarSearchQueryState, filter, telemetryService } = props
-    const hasFilter = filter.length === 0
+    const { onNavbarQueryChange, navbarSearchQueryState, telemetryService } = props
+    const filter = props.filter.trim()
+    const hasFilter = filter.length > 0
 
     const selectedFilters = useMemo(() => {
         if (!hasFilter) {
@@ -650,7 +651,7 @@ const SearchReference = (props: SearchReferenceProps): ReactElement => {
 
     return (
         <div>
-            {props.filter ? (
+            {hasFilter ? (
                 filterList
             ) : (
                 <Tabs index={selectedTab} onChange={setSelectedTab}>